+++ vector/src/main/java/im/vector/app/features/home/room/detail/timeline/action/MessageActionsViewModel.kt
@@ -275,8 +275,8 @@ class MessageActionsViewModel @AssistedInject constructor(@Assisted
                         add(EventSharedAction.ViewEditHistory(informationData))
                     }
 
-                    if (canShare(msgType) && messageContent is MessageWithAttachmentContent) {
-                        add(EventSharedAction.Share(timelineEvent.eventId, messageContent))
+                    if (canShare(msgType)) {
+                        add(EventSharedAction.Share(timelineEvent.eventId, messageContent!!))
                     }
 
                     if (canSave(msgType) && messageContent is MessageWithAttachmentContent) {
@@ -409,6 +409,10 @@ class MessageActionsViewModel @AssistedInject constructor(@Assisted
 
     private fun canShare(msgType: String?): Boolean {
         return when (msgType) {
+            MessageType.MSGTYPE_TEXT,
+            MessageType.MSGTYPE_NOTICE,
+            MessageType.MSGTYPE_EMOTE,
+            MessageType.MSGTYPE_LOCATION,
             MessageType.MSGTYPE_IMAGE,
             MessageType.MSGTYPE_AUDIO,
             MessageType.MSGTYPE_VIDEO,