+++ CHANGES.md
@@ -10,6 +10,7 @@ Improvements 🙌:
  - PIN code: request PIN code if phone has been locked
  - Small optimisation of scrolling experience in timeline (#2114)
  - Allow user to reset cross signing if he has no way to recover (#2052)
+ - Ability to share text
  - Create home shortcut for any room (#1525)
  - Can't confirm email due to killing by Android (#2021)
  - Add a menu item to open the setting in room list and in room (#2171)